cuda - separate jit functionality into separate file
