+++ backends/cuda-shared/ceed-cuda-shared-basis.c
@@ -20,7 +20,7 @@
 #include <cuda_runtime.h>
 #include <stddef.h>
 #include "ceed-cuda-shared.h"
-#include "../cuda/ceed-cuda.h"
+#include "../cuda/ceed-cuda-jit.h"
 #include "kernel-strings/cuda-shared-basis.h"
 
 //------------------------------------------------------------------------------
