+++ backends/cuda/ceed-cuda.c
@@ -19,151 +19,10 @@
 #include <cublas_v2.h>
 #include <cuda.h>
 #include <cuda_runtime.h>
-#include <nvrtc.h>
-#include <stdarg.h>
 #include <stdbool.h>
-#include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
 #include "ceed-cuda.h"
-
-//------------------------------------------------------------------------------
-// Compile CUDA kernel
-//------------------------------------------------------------------------------
-int CeedCompileCuda(Ceed ceed, const char *source, CUmodule *module,
-                    const CeedInt numopts, ...) {
-  int ierr;
-  cudaFree(0); // Make sure a Context exists for nvrtc
-  nvrtcProgram prog;
-  CeedChk_Nvrtc(ceed, nvrtcCreateProgram(&prog, source, NULL, 0, NULL, NULL));
-
-  // Get kernel specific options, such as kernel constants
-  const int optslen = 32;
-  const int optsextra = 4;
-  const char *opts[numopts + optsextra];
-  char buf[numopts][optslen];
-  if (numopts > 0) {
-    va_list args;
-    va_start(args, numopts);
-    char *name;
-    int val;
-    for (int i = 0; i < numopts; i++) {
-      name = va_arg(args, char *);
-      val = va_arg(args, int);
-      snprintf(&buf[i][0], optslen,"-D%s=%d", name, val);
-      opts[i] = &buf[i][0];
-    }
-    va_end(args);
-  }
-
-  // Standard backend options
-  if (CEED_SCALAR_TYPE == CEED_SCALAR_FP32) {
-    opts[numopts]     = "-DCeedScalar=float";
-  } else {
-    opts[numopts]     = "-DCeedScalar=double";
-  }
-  opts[numopts + 1] = "-DCeedInt=int";
-  opts[numopts + 2] = "-default-device";
-  struct cudaDeviceProp prop;
-  Ceed_Cuda *ceed_data;
-  ierr = CeedGetData(ceed, &ceed_data); CeedChkBackend(ierr);
-  ierr = cudaGetDeviceProperties(&prop, ceed_data->deviceId);
-  CeedChk_Cu(ceed, ierr);
-  char buff[optslen];
-  snprintf(buff, optslen,"-arch=compute_%d%d", prop.major, prop.minor);
-  opts[numopts + 3] = buff;
-
-  // Compile kernel
-  nvrtcResult result = nvrtcCompileProgram(prog, numopts + optsextra, opts);
-  if (result != NVRTC_SUCCESS) {
-    size_t logsize;
-    CeedChk_Nvrtc(ceed, nvrtcGetProgramLogSize(prog, &logsize));
-    char *log;
-    ierr = CeedMalloc(logsize, &log); CeedChkBackend(ierr);
-    CeedChk_Nvrtc(ceed, nvrtcGetProgramLog(prog, log));
-    return CeedError(ceed, CEED_ERROR_BACKEND, "%s\n%s",
-                     nvrtcGetErrorString(result), log);
-  }
-
-  size_t ptxsize;
-  CeedChk_Nvrtc(ceed, nvrtcGetPTXSize(prog, &ptxsize));
-  char *ptx;
-  ierr = CeedMalloc(ptxsize, &ptx); CeedChkBackend(ierr);
-  CeedChk_Nvrtc(ceed, nvrtcGetPTX(prog, ptx));
-  CeedChk_Nvrtc(ceed, nvrtcDestroyProgram(&prog));
-
-  CeedChk_Cu(ceed, cuModuleLoadData(module, ptx));
-  ierr = CeedFree(&ptx); CeedChkBackend(ierr);
-  return CEED_ERROR_SUCCESS;
-}
-
-//------------------------------------------------------------------------------
-// Get CUDA kernel
-//------------------------------------------------------------------------------
-int CeedGetKernelCuda(Ceed ceed, CUmodule module, const char *name,
-                      CUfunction *kernel) {
-  CeedChk_Cu(ceed, cuModuleGetFunction(kernel, module, name));
-  return CEED_ERROR_SUCCESS;
-}
-
-// Run kernel with block size selected automatically based on the kernel (which
-// may use enough registers to require a smaller block size than the hardware is
-// capable).
-int CeedRunKernelAutoblockCuda(Ceed ceed, CUfunction kernel, size_t points,
-                               void **args) {
-  int min_grid_size, max_block_size;
-  CeedChk_Cu(ceed, cuOccupancyMaxPotentialBlockSize(&min_grid_size,
-             &max_block_size, kernel, NULL, 0, 0x10000));
-  CeedChkBackend(CeedRunKernelCuda(ceed, kernel, CeedDivUpInt(points,
-                                   max_block_size), max_block_size, args));
-  return 0;
-}
-
-//------------------------------------------------------------------------------
-// Run CUDA kernel
-//------------------------------------------------------------------------------
-int CeedRunKernelCuda(Ceed ceed, CUfunction kernel, const int gridSize,
-                      const int blockSize, void **args) {
-  CUresult result = cuLaunchKernel(kernel, gridSize, 1, 1, blockSize, 1,
-                                   1, 0, NULL, args, NULL);
-  if (result == CUDA_ERROR_LAUNCH_OUT_OF_RESOURCES) {
-    int max_threads_per_block, shared_size_bytes, num_regs;
-    cuFuncGetAttribute(&max_threads_per_block,
-                       CU_FUNC_ATTRIBUTE_MAX_THREADS_PER_BLOCK, kernel);
-    cuFuncGetAttribute(&shared_size_bytes, CU_FUNC_ATTRIBUTE_SHARED_SIZE_BYTES,
-                       kernel);
-    cuFuncGetAttribute(&num_regs, CU_FUNC_ATTRIBUTE_NUM_REGS, kernel);
-    return CeedError(ceed, CEED_ERROR_BACKEND,
-                     "CUDA_ERROR_LAUNCH_OUT_OF_RESOURCES: max_threads_per_block %d on block size (%d,%d,%d), shared_size %d, num_regs %d",
-                     max_threads_per_block, blockSize, 1, 1, shared_size_bytes, num_regs);
-  } else CeedChk_Cu(ceed, result); // NOLINT
-  return CEED_ERROR_SUCCESS;
-}
-
-//------------------------------------------------------------------------------
-// Run CUDA kernel for spatial dimension
-//------------------------------------------------------------------------------
-int CeedRunKernelDimCuda(Ceed ceed, CUfunction kernel, const int gridSize,
-                         const int blockSizeX, const int blockSizeY,
-                         const int blockSizeZ, void **args) {
-  CeedChk_Cu(ceed, cuLaunchKernel(kernel, gridSize, 1, 1,
-                                  blockSizeX, blockSizeY, blockSizeZ,
-                                  0, NULL, args, NULL));
-  return CEED_ERROR_SUCCESS;
-}
-
-//------------------------------------------------------------------------------
-// Run CUDA kernel for spatial dimension with sharde memory
-//------------------------------------------------------------------------------
-int CeedRunKernelDimSharedCuda(Ceed ceed, CUfunction kernel, const int gridSize,
-                               const int blockSizeX, const int blockSizeY,
-                               const int blockSizeZ, const int sharedMemSize,
-                               void **args) {
-  CeedChk_Cu(ceed, cuLaunchKernel(kernel, gridSize, 1, 1,
-                                  blockSizeX, blockSizeY, blockSizeZ,
-                                  sharedMemSize, NULL, args, NULL));
-  return CEED_ERROR_SUCCESS;
-}
 
 //------------------------------------------------------------------------------
 // CUDA preferred MemType
