+++ backends/hip-gen/ceed-hip-gen-operator.c
@@ -19,7 +19,7 @@
 #include <stddef.h>
 #include "ceed-hip-gen.h"
 #include "ceed-hip-gen-operator-build.h"
-#include "../hip/ceed-hip-compile.h"
+#include "../hip/ceed-hip-jit.h"
 
 //------------------------------------------------------------------------------
 // Destroy operator
