+++ backends/hip/ceed-hip-jit.h
@@ -14,8 +14,8 @@
 // software, applications, hardware, advanced system engineering and early
 // testbed platforms, in support of the nation's exascale computing imperative.
 
-#ifndef _ceed_hip_compile_h
-#define _ceed_hip_compile_h
+#ifndef _ceed_hip_jit_h
+#define _ceed_hip_jit_h
 
 #include <ceed/ceed.h>
 #include <hip/hip_runtime.h>
@@ -42,4 +42,4 @@ CEED_INTERN int CeedRunKernelDimSharedHip(Ceed ceed, hipFunction_t kernel,
     const int blockSizeY, const int blockSizeZ,
     const int sharedMemSize, void **args);
 
-#endif // _ceed_hip_compile_h
+#endif // _ceed_hip_jit_h
